pyproject.toml
[tool.poetry]
name = "duckbot"
version = "0.0.0"
description = ""
authors = ["Shuri2060 <29679284+Shuri2060@users.noreply.github.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
aiohttp = "^3.9.5"
aiosignal = "^1.3.1"
alembic = "^1.13.2"
annotated-types = "^0.7.0"
async-timeout = "^4.0.3"
eth-account = "^0.13.3"
eth-hash = "^0.7.0"
eth-keyfile = "^0.8.1"
eth-keys = "^0.5.1"
eth-rlp = "^2.1.0"
eth-typing = "^5.0.0"
eth-utils = "^5.0.0"
eth_abi = "^5.1.0"
mnemonic = "^0.21"
netaddr = "^1.3.0"
numpy = "^2.1.1"
parsimonious = "^0.10.0"
python-telegram-bot = "^21.2"
redis = "^5.0.8"
requests = "^2.32.2"
SQLAlchemy = "^2.0.35"
web3 = "^6.0.0b8"
websockets = "^13.0.1"
yarl = "^1.11.1"
jupyterlab = "^4.0.10"
psycopg2  = "^2.9.9"
psycopg2-binary = "^2.9.9"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
